Stamina now depletes with actions, delay then regens at a set rate.

+++ SyntyTestProject/Assets/HealthBar.cs
@@ -6,11 +6,11 @@ using UnityEngine.UI;
 public class HealthBar : MonoBehaviour
 {
     public Slider slider;
-    public void SetMaxHealth(int maxHealth) {
+    public void SetMaxHealth(float maxHealth) {
         slider.maxValue = maxHealth;
         slider.value = maxHealth;
     }
-    public void SetCurrentHealth(int currentHealth) {
+    public void SetCurrentHealth(float currentHealth) {
         slider.value = currentHealth;
     }
 }
